fix lossless scaling profile limits

+++ src/process.h
@@ -139,7 +139,7 @@ namespace proc {
     bool gen2_framegen_fix;
     bool lossless_scaling_framegen;
     std::string frame_generation_provider {"lossless-scaling"};
-    std::optional<int> lossless_scaling_target_fps;
+    std::optional<double> lossless_scaling_target_fps;
     std::optional<int> lossless_scaling_rtss_limit;
     std::string lossless_scaling_profile {"custom"};
     lossless_scaling_profile_overrides_t lossless_scaling_recommended;
