+++ src/stream.h
@@ -42,7 +42,7 @@ namespace stream {
     bool gen2_framegen_fix;
     bool lossless_scaling_framegen;
     std::string frame_generation_provider;
-    std::optional<int> lossless_scaling_target_fps;
+    std::optional<double> lossless_scaling_target_fps;
     std::optional<int> lossless_scaling_rtss_limit;
   };
